requirements.txt
python-docx
PyPDF2
discord.py>=2.4.0
openai
python-dotenv>=1.0.1
aiohttp>=3.10.4
flask>=3.0.3
waitress>=3.0.0
anthropic>=0.34.0
httpx>=0.27.0
requests>=2.32.3
google-generativeai>=0.8.3
google-ai-generativelanguage==0.6.10
yt-dlp>=2024.12.13
mysql-connector-python>=9.1.0
colorama>=0.4.6
aiofiles>=24.1.0
pynacl>=1.5.0
pandas
